+++ src/lib/Zikula/Bundle/CoreInstallerBundle/Form/Type/AbstractType.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Zikula\Bundle\CoreInstallerBundle\Form;
+namespace Zikula\Bundle\CoreInstallerBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType as BaseAbstractType;
 use Zikula\Common\Translator\TranslatorTrait;
