+++ src/lib/Zikula/Bundle/CoreInstallerBundle/Form/Type/CreateAdminType.php
@@ -21,7 +21,6 @@ use Symfony\Component\Validator\Constraints\Email;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\Regex;
-use Zikula\Bundle\CoreInstallerBundle\Form\AbstractType;
 use Zikula\Common\Translator\IdentityTranslator;
 use Zikula\UsersModule\Constant as UsersConstant;
