+++ src/system/UsersModule/Form/Type/ConfigType/AuthenticationMethodsType.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Zikula\UsersModule\Form\ConfigType;
+namespace Zikula\UsersModule\Form\Type\ConfigType;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
